requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
yfinance>=0.2.0
pytest>=7.0.0
pandera>=0.15.0
